refactor: service crud

+++ backend/src/use-cases/delete-service.ts
@@ -1,4 +1,5 @@
-import { ServicesRepository } from '@/repositories/services-repository';
+import { ServicesRepository } from "@/repositories/services-repository";
+import { ResourceNotFoundError } from "./errors/resource-not-found-error";
 
 interface DeleteServiceUseCaseRequest {
   id: string;
@@ -7,9 +8,13 @@ interface DeleteServiceUseCaseRequest {
 export class DeleteServiceUseCase {
   constructor(private servicesRepository: ServicesRepository) {}
 
-  async execute({
-    id
-  }: DeleteServiceUseCaseRequest): Promise<void> {
+  async execute({ id }: DeleteServiceUseCaseRequest): Promise<void> {
+    const service = this.servicesRepository.findById(id);
+
+    if (!service) {
+      throw new ResourceNotFoundError();
+    }
+
     await this.servicesRepository.delete(id);
 
     return;
